Converted nnz to 8-byte integer.

+++ src/inversion/inverse_problem.F90
@@ -76,7 +76,7 @@ subroutine inversion_initialize(this, par, myrank)
   this%nelements = par%nelements
 
   call this%matrix%initialize(par%ndata(1) + par%nelements_total, &
-                              par%ndata(1) * par%nelements + par%nelements, myrank)
+                              int8(par%ndata(1) * par%nelements + par%nelements), myrank)
 
   ierr = 0
