+++ src/tests/tests_inversion.f90
@@ -15,7 +15,7 @@
 !===============================================================================================
 ! Unit tests for inversion part.
 !
-! Author: Vitaliy Ogarko, UWA, CET, Australia, 2015-2016.
+! Author: Vitaliy Ogarko, UWA, CET, Australia.
 !===============================================================================================
 module tests_inversion
 
@@ -95,7 +95,7 @@ subroutine test_add_damping_identity_matrix(myrank, nbproc)
   enddo
 
   call isensit%initialize(par%ndata(1) + par%nelements_total, &
-                          par%ndata(1) * par%nelements + par%nelements, myrank)
+                          int8(par%ndata(1) * par%nelements + par%nelements), myrank)
 
   call damping%initialize(par%nelements, par%alpha(1), par%problem_weight(1), par%norm_power, &
                           par%compression_type, par%nx, par%ny, par%nz)
@@ -214,12 +214,12 @@ subroutine test_cross_gradient_calculate(myrank, nbproc, derivative_type)
 
   call cross_grad%initialize(nx, ny, nz, nelements, myrank)
 
-  call matrix%initialize(3 * nelements_total, cross_grad%get_num_elements(derivative_type), myrank)
+  call matrix%initialize(3 * nelements_total, int8(cross_grad%get_num_elements(derivative_type)), myrank)
 
   call cross_grad%calculate(model1, model2, column_weight1, column_weight2, &
                             matrix, b_RHS, .true., derivative_type, myrank, nbproc)
 
-  matrix_nel_loc = matrix%get_number_elements()
+  matrix_nel_loc = int(matrix%get_number_elements())
 
   call mpi_allreduce(matrix_nel_loc, matrix_nel_glob, 1, MPI_INTEGER, MPI_SUM, MPI_COMM_WORLD, ierr)
